[New] `nvm uninstall`: Check installation dir permissions before uninstalling

+++ nvm.sh
@@ -1828,6 +1828,27 @@ nvm() {
         NVM_PREFIX="$(nvm_node_prefix)"
         NVM_SUCCESS_MSG="Uninstalled node $VERSION"
       fi
+
+      # Check version dir permissions
+      local PERMISSIONS_OK=true
+      check_file_permissions() {
+        for FILE in $1/* $1/.[!.]* $1/..?* ; do
+            if [ -d "$FILE" ]; then
+              check_file_permissions "$FILE"
+            elif [ -e "$FILE" ]; then
+              [ ! -w "$FILE" ] && PERMISSIONS_OK=false
+            fi
+        done
+      }
+      check_file_permissions "$VERSION_PATH"
+      if [ "$PERMISSIONS_OK" = false ]; then
+        >&2 echo 'Cannot uninstall, incorrect permissions on installation folder.'
+        >&2 echo 'This is usually caused by running `npm install -g` as root. Run the following command as root to fix the permissions and then try again.'
+        >&2 echo
+        >&2 echo "  chown -R $(whoami) \"$VERSION_PATH\""
+        return 1
+      fi
+
       # Delete all files related to target version.
       command rm -rf "$NVM_DIR/src/$NVM_PREFIX-$VERSION" \
              "$NVM_DIR/src/$NVM_PREFIX-$VERSION.tar.*" \
